generalize MERF to MixedEffectsModel with separate child classes for MERF and MEEBM
